Fixes and cleanups related to inlines
This is a set of fixes and cleanups related to inlines. These were
discovered while testing inlines and getting them working
correctly. This includes:

* Update `doBindDismiss` calls to use the new function name,
`_bindDismiss`. This also removes an extra call in the
constructor. The bind dismiss is already being called in the
`bindElem` which is called by the constructor.
* Support multiple dismiss elems. This updates the logic to look
for multiple dismiss elems and bind to their click events. This
also checks all of the dismiss elems to see if any of them are
configured with a confirm data attr or are disabled when handling
their click.