(hotfix) jux onkey delay ms config with other onkey configs
Previously, this was jux'd next to the event handler b/c it used
to be a variable before we removed the manual onkey timeout logic.
Now that it isn't needed directly in the event handler logic, this
setting should be juxed with the other onkey settings for the
filter elem.