Hotfix: Fix passing strings to `Romo.array`
This fixes passing strings to `Romo.array` and having it
properly put them into an array. The logic was previously
splitting the string into an array of its characters. This is
because strings are array like but the `Romo.array` has a check
to specifically avoid it. The issue was it was using the `Object`
wrapped string which returns "object" for its `typeof` instead
of "string". This switches it to use original value that was
passed in to check if its a string or not.