(hotfix) update base `parents` meth to work with the body elem
Previously this would fail b/c body elem's parent is the document
and it isn't a typical element.  Plus this method is designed to
only return elements up to (but not including) the document.  This
tests for this case and returns and empty array if called on the
body elem.

I noticed this while debugging errors in the dropdown window body
click event handlers (which were calling `parents` on the body
elem).