base: update `array` to work with select elems
Select elems, it turns out, are "array like" in that they have a
length attr.  This caused them to return an array of their options
when passed to the prev `array` implementation.

This updates the logic to specifically test for select elems and
handle them much like functions are specifically tested for.  I
think these are the only elems that are "array like".