(hotfix) don't call init ui callbacks in timeouts
I originally added them to call like this to emulate the old
event implementation we used in the jQuery based Romo.  However,
this calls the init callbacks to be called out of order in that
they will be called after all of the `Romo.ready` callbacks.  This
causes some components to break b/c they expect the init callbacks
to run in order of how they are defined.

This updates the callbacks to just be run immediately, without any
timeouts.