(hotfix) base: rename remove node vars in parent-child elems code
These code deals with nodes (not necessarily elems) so this
renames the variables related to the removed nodes to "nodes".
This is a minor cosmetic change but I want to be clear about what
the vars represent and how they should be treated.

Note: I chose to keep "elem" in the `remove` logic var names b/c
the only type of nodes this processes are elems.  It should be
thought of as nodes since non elem nodes could be passed and the
node type is checked at the beginning.  But after that check, the
nodes should be considered element nodes.