base: rework elems init API
We noticed, while evaluating updates to the api to task optional
elem collections, that the API to setup auto elem init components
was a bit cumbersome and involved a bit of boilerplate code.  This
reworks the API to make it so you only make a single api call
passing in the elem selector and the component class.  Romo now
handles all of the boilerplate init logic in it's private helpers.

This also renames the api, in general, as "elems init".  "UI" was
a little ambiguous and components may or may not be related to UI.
A better way to describe all of this is initializing elems with a
component.  The private helpers were also reworked to be slightly
more efficient now that the `find` method can take a set of elems.

Note: this also removes the `_addEventCallback` method.  This
method isn't used and should have been removed earlier but was
missed.