Update `on`, `off, and `trigger` to take multiple elems
This updates the `on`, `off`, and `trigger` helpers to take
multiple elems instead of just a single elem. This is part of an
effort to make the helpers more convenient when dealing with
collections of elems and avoid having extra noisy `forEach` loops.

This updates the `on` helper to build a single proxy function and
bind it to the multiple elems. This avoids building a proxy
function for each elem but otherwise just loops over the previous
behavior in a `forEach` loop.

This also updates the `trigger` helper to build a single event and
then loop over the passed elems dispatch it. Similar to the `on`
helper this avoids building a new identical event for each elem
which is hopefully slightly more performant.

The also updates the `off` helper but it simply loops over its
elems and runs its previous logic.