Update `IndicatorTextInput` to pass multi elems to romo helpers
This updates the `IndicatorTextInput` component to make use of
romo's helpers that now take multiple elems. This involved adding,
removing, showing, and hiding the elem and indicator elem. This
was the only logic that could make use of the helper function
updates.

This also removes some private functions for showing or hiding
elems since they were just simple wrappers to calling romo's
`show` and `hide`.