Update init replace methods to return single elem
This updates the `initReplace` and `initReplaceHtml` methods to
be consistent with their non init versions and return the single
elem that was replaced. This was changed in PR 227. The previous
helper `triggerInitUI` would cast whatever it was passed as an
array and then iterate over it. It would return the original value
not the array. The new implementation does the same only it
returns the array instead of the original value. This is because
the `_elemsInitTrigger` is passed the array and returns it.

This fixes the problem by having the `initReplace` and
`initReplaceHtml` simply pull the first item out of the array
results that `initElems` returns. This keeps them consistent
with the other replace methods.