(hotfix) dropdown/modal/tooltip: place popup immediately and in timeout
This is to help avoid "flashes" of popups positioning when opened
in these components.

We originally put the popup placement calls in the timeout to
delay positioning until all component in the popup had a chance to
mutate the UI and get initialized.  This is good for getting the
ultimate position correct.  However, delaying like this causes the
popup to be positioned incorrectly until this is called.

This means the UI "flashes" the incorrect position and immediately
corrects it (but the flash is very noticeable).  This change means
the popup is at least somewhat positioned correctly while we wait
on the final positioning which avoids the "flash".