(hotfix) properly return event function return values
This is a bug from switching to the event function API in PR 239.
I added an event proxy function to get unique event functions per
component instance.  However, I forgot to return the proxy return
values.  This caused any prevent default and stop propagation
directives added by `return false` to not get added.  This was
causing weird event handling.

I noticed this using selects.  I would hit enter on an option to
select it and the form the select was in would submit.  I've
added a couple minor updates to the option list dropdown I noticed
while debugging this issue.