Hotfix: Handle removing elems with no parent
This updates the `Romo.remove` helper to handle when its passed
an elem that doesn't have a parent (is already detached from the
DOM). In this case the `remove` method ignores the elem since
it's already been removed from the DOM. This fixes an issue with
the `RomoParentChildElems` trying to remove elems that have
already been removed.

This was noticed when replacing an elem that was added to the
`RomoParentChildElems`, in this case a picker input. The parent
elem of the picker input had called `updateHtml` with new content
which causes the picker input elem to be detached from the DOM
(this is because `updateHtml` sets the parent elems inner HTML to
an empty string). The `RomoParentChildElems` saw the change and
told the picker input to remove its child elems. One of it's
"child" elems (via `RomoParentChildElems`) is the romo picker
wrapper which is under the same parent as the picker input and
had also already been detached from the DOM. This caused the
`remove` function to error trying to remove the romo picker
wrapper because it was already detached.

Note: This updates the `remove` function to return the elems it
removed even if it doesn't actually detach them. This is
consistent with it's previous return value because the
`removeChild` function returned the elem that was removed.