(hotfix) some cleanups after the popup stack rework
This moves the private open/close popup methods used by the popup
stack to jux them in our private methods sections on the modal
and dropdown components.  This should have been done when we
added the popup stack but I didn't want the noise of the move
in the diff.

This also adds position relative on the modal body/content.  This
was noticed while researching other ways to place popups that we
didn't end up using.  However, this is the spirit of the
component and should have been in place from the start.