(hotfix) tooltip: allow adding content with incomplete html
This switches to wrapping the given content in a div tag and
always using `updateHtml` to set the content.  This allows
show both plain text content and text content with partial html
like br tags.  I chose to wrap using a generic div tag (instead
of a span tag) so that setting more complex or non-inline html.