Hotfix: Fix `Romo.elems` working on Internet Explorer
This fixes the `Romo.elems` method working on Internet Explorer.
This is part of an effort to have romo work for later versions of
IE. The `createHTMLDocument` method takes a title argument that is
optional in the other browsers but is required in IE. This simply
switches to passing an empty string which keeps IE from erroring.