Hotfix: Fix `RomoParentChildElems` in Internet Explorer
This fixes the `RomoParentChildElems` logic in Internet Explorer.
This is part of an effort to have romo work for later versions of
IE. The `removedNodes` method returns a `NodeList` object but in
IE the `NodeList` objects don't respond to `forEach`. This fixes
it by converting the `NodeList` into an array using `Romo.array`.
This is consistent with how we handle other `NodeList` values in
`Romo.f` and `Romo.children` and ensures that `forEach` can be
called on the value.